Changed folder structure (pages)
