+++ client/lib/api/book.dart
