+++ client/lib/api/graphql_config.dart
