+++ client/lib/pages/login/email_login_page.dart
@@ -1,27 +1,25 @@
+import 'package:ALPapp/pages/core/auth_page.dart';
 import 'package:ALPapp/services/auth_service.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'index.dart';
-
-class EmailSignUp extends StatefulWidget {
+class EmailLogIn extends StatefulWidget {
   @override
-  _EmailSignUpState createState() => _EmailSignUpState();
+  _EmailLogInState createState() => _EmailLogInState();
 }
 
-class _EmailSignUpState extends State<EmailSignUp> {
-  bool isLoading = false;
+class _EmailLogInState extends State<EmailLogIn> {
   final _formKey = GlobalKey<FormState>();
   TextEditingController emailController = TextEditingController();
-  TextEditingController nameController = TextEditingController();
   TextEditingController passwordController = TextEditingController();
-  TextEditingController ageController = TextEditingController();
+
+  bool isLoading = false;
 
   @override
   Widget build(BuildContext context) {
     var authService = Provider.of<AuthService>(context);
     return Scaffold(
-        appBar: AppBar(title: Text("Sign Up")),
+        appBar: AppBar(title: Text("Login")),
         body: Form(
             key: _formKey,
             child: SingleChildScrollView(
@@ -29,28 +27,9 @@ class _EmailSignUpState extends State<EmailSignUp> {
               Padding(
                 padding: EdgeInsets.all(20.0),
                 child: TextFormField(
-                  controller: nameController,
-                  decoration: InputDecoration(
-                    labelText: "Enter User Name",
-                    enabledBorder: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(10.0),
-                    ),
-                  ),
-                  // The validator receives the text that the user has entered.
-                  validator: (value) {
-                    if (value.isEmpty) {
-                      return 'Enter User Name';
-                    }
-                    return null;
-                  },
-                ),
-              ),
-              Padding(
-                padding: EdgeInsets.all(20.0),
-                child: TextFormField(
                   controller: emailController,
                   decoration: InputDecoration(
-                    labelText: "Enter Email",
+                    labelText: "Enter Email Address",
                     enabledBorder: OutlineInputBorder(
                       borderRadius: BorderRadius.circular(10.0),
                     ),
@@ -58,9 +37,9 @@ class _EmailSignUpState extends State<EmailSignUp> {
                   // The validator receives the text that the user has entered.
                   validator: (value) {
                     if (value.isEmpty) {
-                      return 'Enter an Email Address';
+                      return 'Enter Email Address';
                     } else if (!value.contains('@')) {
-                      return 'Please enter a valid email address';
+                      return 'Please enter a valid email address!';
                     }
                     return null;
                   },
@@ -93,29 +72,17 @@ class _EmailSignUpState extends State<EmailSignUp> {
                 child: RaisedButton(
                   color: Colors.lightBlue,
                   onPressed: () {
-                    //TODO Add try catch implementation
-                    authService.signUpWithMail(
+                    authService.signInWithMail(
                         email: emailController.text,
-                        password: passwordController.text,
-                        userName: nameController.text);
-                    Navigator.of(context).pushAndRemoveUntil(MaterialPageRoute(
-                      builder: (context) {
-                        return IndexPage();
-                      },
-                    ), (route) => false);
+                        password: passwordController.text);
+
+                    Navigator.of(context).pushAndRemoveUntil(
+                        MaterialPageRoute(builder: (context) => AuthPage()),
+                        (route) => false);
                   },
                   child: Text('Submit'),
                 ),
               )
             ]))));
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    nameController.dispose();
-    emailController.dispose();
-    passwordController.dispose();
-    ageController.dispose();
   }
 }
