+++ client/lib/pages/signup/email_signup.dart
@@ -1,25 +1,27 @@
-import 'package:ALPapp/pages/auth_page.dart';
 import 'package:ALPapp/services/auth_service.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-class EmailLogIn extends StatefulWidget {
+import 'package:ALPapp/pages/home/home_page.dart';
+
+class EmailSignUp extends StatefulWidget {
   @override
-  _EmailLogInState createState() => _EmailLogInState();
+  _EmailSignUpState createState() => _EmailSignUpState();
 }
 
-class _EmailLogInState extends State<EmailLogIn> {
+class _EmailSignUpState extends State<EmailSignUp> {
+  bool isLoading = false;
   final _formKey = GlobalKey<FormState>();
   TextEditingController emailController = TextEditingController();
+  TextEditingController nameController = TextEditingController();
   TextEditingController passwordController = TextEditingController();
-
-  bool isLoading = false;
+  TextEditingController ageController = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
     var authService = Provider.of<AuthService>(context);
     return Scaffold(
-        appBar: AppBar(title: Text("Login")),
+        appBar: AppBar(title: Text("Sign Up")),
         body: Form(
             key: _formKey,
             child: SingleChildScrollView(
@@ -27,9 +29,9 @@ class _EmailLogInState extends State<EmailLogIn> {
               Padding(
                 padding: EdgeInsets.all(20.0),
                 child: TextFormField(
-                  controller: emailController,
+                  controller: nameController,
                   decoration: InputDecoration(
-                    labelText: "Enter Email Address",
+                    labelText: "Enter User Name",
                     enabledBorder: OutlineInputBorder(
                       borderRadius: BorderRadius.circular(10.0),
                     ),
@@ -37,9 +39,28 @@ class _EmailLogInState extends State<EmailLogIn> {
                   // The validator receives the text that the user has entered.
                   validator: (value) {
                     if (value.isEmpty) {
-                      return 'Enter Email Address';
+                      return 'Enter User Name';
+                    }
+                    return null;
+                  },
+                ),
+              ),
+              Padding(
+                padding: EdgeInsets.all(20.0),
+                child: TextFormField(
+                  controller: emailController,
+                  decoration: InputDecoration(
+                    labelText: "Enter Email",
+                    enabledBorder: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(10.0),
+                    ),
+                  ),
+                  // The validator receives the text that the user has entered.
+                  validator: (value) {
+                    if (value.isEmpty) {
+                      return 'Enter an Email Address';
                     } else if (!value.contains('@')) {
-                      return 'Please enter a valid email address!';
+                      return 'Please enter a valid email address';
                     }
                     return null;
                   },
@@ -72,17 +93,29 @@ class _EmailLogInState extends State<EmailLogIn> {
                 child: RaisedButton(
                   color: Colors.lightBlue,
                   onPressed: () {
-                    authService.signInWithMail(
+                    //TODO Add try catch implementation
+                    authService.signUpWithMail(
                         email: emailController.text,
-                        password: passwordController.text);
-
-                    Navigator.of(context).pushAndRemoveUntil(
-                        MaterialPageRoute(builder: (context) => AuthPage()),
-                        (route) => false);
+                        password: passwordController.text,
+                        userName: nameController.text);
+                    Navigator.of(context).pushAndRemoveUntil(MaterialPageRoute(
+                      builder: (context) {
+                        return IndexPage();
+                      },
+                    ), (route) => false);
                   },
                   child: Text('Submit'),
                 ),
               )
             ]))));
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    nameController.dispose();
+    emailController.dispose();
+    passwordController.dispose();
+    ageController.dispose();
   }
 }
