+++ client/lib/pages/util_pages/book_details_page.dart
